- added an id field to RDRCase and corresponding change to Serializer - added CornerstoneView component to show when building a rule - cosmetic changes to the CaseView table to use MUI and reduce the padding

+++ src/jvmIntegrationTest/kotlin/io/rippledown/integration/proxy/ServerProxy.kt
@@ -2,6 +2,7 @@ package io.rippledown.integration.proxy
 
 import io.rippledown.server.STARTING_SERVER
 import io.rippledown.integration.restclient.RESTClient
+import io.rippledown.server.IN_MEMORY
 import org.awaitility.Awaitility.await
 import java.io.File
 import java.nio.file.Paths
@@ -25,7 +26,8 @@ class ServerProxy {
         findJar()
         dirProxy.createAndCleanManagedDirectories()
         logProxy.deleteLogfiles()
-        process = ProcessBuilder("java", "-jar", jarFile.absolutePath)
+//        process = ProcessBuilder("java",  "-jar", jarFile.absolutePath)
+        process = ProcessBuilder("java",  "-jar", jarFile.absolutePath, IN_MEMORY)
             .redirectErrorStream(true)
             .redirectOutput(systemOutputFile)
             .directory(dirProxy.tempDir())
@@ -50,6 +52,8 @@ class ServerProxy {
 
     fun shutdown() {
         restClient.shutdown()
-        process.destroyForcibly().waitFor()
+        if ( this::process.isInitialized) {
+            process.destroyForcibly().waitFor()
+        }
     }
 }
